feat: add the first migration
