+++ alembic.ini
@@ -39,7 +39,7 @@ prepend_sys_path = .
 # are written from script.py.mako
 # output_encoding = utf-8
 
-sqlalchemy.url = driver://user:pass@localhost/dbname
+sqlalchemy.url = mysql+pymysql://root@localhost/huahua
 
 
 [post_write_hooks]
