+++ alembic/env.py
@@ -1,3 +1,4 @@
+from model import Base
 from logging.config import fileConfig
 
 from sqlalchemy import engine_from_config
@@ -17,7 +18,7 @@ fileConfig(config.config_file_name)
 # for 'autogenerate' support
 # from myapp import mymodel
 # target_metadata = mymodel.Base.metadata
-target_metadata = None
+target_metadata = Base.metadata
 
 # other values from the config, defined by the needs of env.py,
 # can be acquired:
