+++ alembic/versions/301f8f1e4c66_create_servers_table.py
@@ -1,0 +1,37 @@
+"""create servers table
+
+Revision ID: 301f8f1e4c66
+Revises: 
+Create Date: 2021-08-14 09:28:41.868763
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '301f8f1e4c66'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('servers',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('address', sa.VARCHAR(length=512), nullable=True),
+    sa.Column('alias', sa.VARCHAR(length=256), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_servers_address'), 'servers', ['address'], unique=False)
+    op.create_index(op.f('ix_servers_alias'), 'servers', ['alias'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_servers_alias'), table_name='servers')
+    op.drop_index(op.f('ix_servers_address'), table_name='servers')
+    op.drop_table('servers')
+    # ### end Alembic commands ###
